# apifirst-api/openapi/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: API First with Spring Boot
  termsOfService: https://springframework.guru/
  contact:
    name: Contact our support
    email: dboeckli@gmail.com
    url: https://springframework.guru/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  description: >
    This is an **example** API to demonstrate features of the OpenAPI
    specification.
...
servers:
  - url: https://{tenant}/api/v1
    variables:
      tenant:
        default: www
        description: Your tenant id
  - url: https://springframework.guru/api/v1
...
paths:
...
  /v1/products:
    get:
      tags:
        - Product
      summary: Get a list of products
      description: Returns a list of products
      operationId: listProducts
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/product_list'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags:
        - Product
      summary: Create a new product
      description: Creates a new product
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/product_create'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: The URL of the newly created product
              schema:
                type: string
                example: /v1/products/27ca5995-f6e1-46bb-8f18-52458b7ed96e
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/products/{productId}:
    get:
      tags:
        - Product
      summary: Get a product by ID
      description: Get a product by ID
      operationId: getProductById
      parameters:
        - $ref: '#/components/parameters/product_id'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/product'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      tags:
        - Product
      summary: Update a product by ID
      description: |
        Update a product by ID
      operationId: updateProductById
      parameters:
        - $ref: '#/components/parameters/product_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/product_update'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/product'
        '400':
          description: Bad Request
        '404':
          description: Not Found
    patch:
      tags:
        - Product
      summary: Patch a product by ID
      description: |
        Patch a product by ID
      operationId: patchProductById
      parameters:
        - $ref: '#/components/parameters/product_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/product_patch'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/product'
        '400':
          description: Bad Request
        '404':
          description: Not Found
    delete:
      tags:
        - Product
      summary: Delete a Product by ID
      description: |
        Delete a Product by ID
      operationId: deleteProductById
      parameters:
        - $ref: '#/components/parameters/product_id'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '404':
          description: Not Found
...
security:
  - basic_auth: []
components:
  securitySchemes:
    main_auth:
      description: Example description text of the OAuth2 scheme.
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://example.com/api/oauth/dialog
          scopes:
            read:users: read user info
            write:users: modify or remove users
    api_key:
      description: Example description text of the API key scheme.
      type: apiKey
      in: header
      name: api_key
    basic_auth:
      type: http
      scheme: basic
  schemas:
    uuid_id:
      type: string
      format: uuid
      description: System generated unique identifier for the resource.
      readOnly: true
      nullable: true
...
    iso_timestamp:
      type: string
      format: date-time
      description: >-
        ISO 8601 Timestamp. Date Created and Date Updated are system managed
        values, should not be sent in request payloads - will be ignored.
      example: '2022-08-22T11:43:06-07:00'
      title: ISO Timestamp
      readOnly: true
      nullable: true
...
    Problem:
      type: object
      additionalProperties: true
      minProperties: 1
      description: >-
        The Problem Details JSON Object
        [[RFC7807](https://tools.ietf.org/html/rfc7807)].
      properties:
        type:
          type: string
          description: >-
            A URI reference [[RFC3986](https://tools.ietf.org/html/rfc3986)]
            that identifies the problem type. It should provide human-readable
            documentation for the problem type. When this member is not present,
            its value is assumed to be "about:blank".
          format: uri
        title:
          type: string
          description: >-
            A short, human-readable summary of the problem type. It SHOULD NOT
            change from occurrence to occurrence of the problem, except for
            purposes of localization.
        status:
          type: integer
          description: The HTTP status code.
          minimum: 400
          maximum: 599
        detail:
          type: string
          description: >-
            A human-readable explanation specific to this occurrence of the
            problem.
        instance:
          type: string
          description: >-
            A URI reference that identifies the specific occurrence of the
            problem.  It may or may not yield further information if
            dereferenced.
...
    dimensions:
      type: object
      properties:
        length:
          type: integer
          description: The length of the product
          minimum: 1
          maximum: 999
          example: 10
        width:
          type: integer
          description: The width of the product
          minimum: 1
          maximum: 999
          example: 10
        height:
          type: integer
          description: The height of the product
          minimum: 1
          maximum: 999
          example: 10
      description: The dimensions of the product
      required:
        - length
        - width
        - height
    category:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuid_id'
        category:
          type: string
          nullable: false
          description: The category of the product
          minLength: 3
          maxLength: 25
          example: Electronics
        description:
          type: string
          description: The description of the category
          minLength: 3
          maxLength: 255
          example: Electronics and Gadgets
        category_code:
          type: string
          description: The category code of the product
          format: alphanumeric
          minLength: 3
          maxLength: 25
          example: ELECTRONICS
        dateCreated:
          $ref: '#/components/schemas/iso_timestamp'
        dateUpdated:
          $ref: '#/components/schemas/iso_timestamp'
      required:
        - category
        - description
        - category_code
    image:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuid_id'
        url:
          type: string
          format: url
          description: The url of the image
          minLength: 5
          maxLength: 500
          example: >-
            https://springframework.guru/wp-content/uploads/2019/08/spring-framework-guru-logo.jpg
        altText:
          type: string
          nullable: true
          minLength: 3
          maxLength: 255
        dateCreated:
          $ref: '#/components/schemas/iso_timestamp'
        dateUpdated:
          $ref: '#/components/schemas/iso_timestamp'
      required:
        - url
    product:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuid_id'
        description:
          type: string
          description: The description of the product
          minLength: 3
          maxLength: 255
        dimensions:
          $ref: '#/components/schemas/dimensions'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/category'
        images:
          type: array
          items:
            $ref: '#/components/schemas/image'
        price:
          type: string
          pattern: ^-?(?:0|[1-9]\d{0,2}(?:,?\d{3})*)(?:\.\d+)?$
        cost:
          type: string
          pattern: ^-?(?:0|[1-9]\d{0,2}(?:,?\d{3})*)(?:\.\d+)?$
        dateCreated:
          $ref: '#/components/schemas/iso_timestamp'
        dateUpdated:
          $ref: '#/components/schemas/iso_timestamp'
      required:
        - description
        - price
    product_list:
      type: array
      items:
        $ref: '#/components/schemas/product'
    product_create:
      type: object
      properties:
        description:
          type: string
          description: The description of the product
          minLength: 3
          maxLength: 255
        dimensions:
          $ref: '#/components/schemas/dimensions'
        categories:
          type: array
          description: The category codes for the product
          items:
            type: string
        images:
          type: array
          items:
            $ref: '#/components/schemas/image'
        price:
          type: string
          pattern: ^-?(?:0|[1-9]\d{0,2}(?:,?\d{3})*)(?:\.\d+)?$
        cost:
          type: string
          pattern: ^-?(?:0|[1-9]\d{0,2}(?:,?\d{3})*)(?:\.\d+)?$
      required:
        - description
        - price
    product_image_update:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: System generated unique identifier for the image.
          nullable: true
        url:
          type: string
          format: url
          description: The url of the image
          minLength: 5
          maxLength: 500
          example: >-
            https://springframework.guru/wp-content/uploads/2019/08/spring-framework-guru-logo.jpg
        altText:
          type: string
          nullable: true
          minLength: 3
          maxLength: 255
      required:
        - url
        - id
    product_update:
      type: object
      properties:
        description:
          type: string
          description: The description of the product
          minLength: 3
          maxLength: 255
        dimensions:
          $ref: '#/components/schemas/dimensions'
        categories:
          type: array
          description: The category codes for the product
          items:
            type: string
        images:
          type: array
          items:
            $ref: '#/components/schemas/product_image_update'
        price:
          type: string
          pattern: ^-?(?:0|[1-9]\d{0,2}(?:,?\d{3})*)(?:\.\d+)?$
        cost:
          type: string
          pattern: ^-?(?:0|[1-9]\d{0,2}(?:,?\d{3})*)(?:\.\d+)?$
      required:
        - description
        - price
    product_dimensions_patch:
      type: object
      properties:
        length:
          type: integer
          description: The length of the product
          minimum: 1
          maximum: 999
          example: 10
          nullable: true
        width:
          type: integer
          description: The width of the product
          minimum: 1
          maximum: 999
          example: 10
          nullable: true
        height:
          type: integer
          description: The height of the product
          minimum: 1
          maximum: 999
          example: 10
          nullable: true
      description: The dimensions of the product
    product_image_patch:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the image
        url:
          type: string
          format: url
          description: The url of the image
          minLength: 5
          maxLength: 500
          nullable: true
          example: >-
            https://springframework.guru/wp-content/uploads/2019/08/spring-framework-guru-logo.jpg
        altText:
          type: string
          nullable: true
          minLength: 3
          maxLength: 255
      required:
        - id
    product_patch:
      type: object
      properties:
        description:
          type: string
          description: The description of the product
          minLength: 3
          maxLength: 255
          nullable: true
        dimensions:
          $ref: '#/components/schemas/product_dimensions_patch'
        categories:
          type: array
          description: The category codes for the product
          items:
            type: string
          nullable: true
        images:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/product_image_patch'
        price:
          type: string
          pattern: ^-?(?:0|[1-9]\d{0,2}(?:,?\d{3})*)(?:\.\d+)?$
          nullable: true
        cost:
          type: string
          pattern: ^-?(?:0|[1-9]\d{0,2}(?:,?\d{3})*)(?:\.\d+)?$
          nullable: true
...
  parameters:
...
    product_id:
      name: productId
      in: path
      description: The ID of the product to retrieve
      required: true
      schema:
        type: string
        format: uuid
...